flask/flask-intro-nov-25-exercise-fromDisco/requirements.txt
Flask
black
psycopg2-binary
ipython
mock
pytest